icons, design, moved game logic (#15, #18, #20)
