+++ isl/src/index.css
@@ -58,6 +58,7 @@ body {
   background-color: #BDD3FC;
   border: 0;
   padding: 1vmin;
+  font-size: 1.25em;
 }
 
 
@@ -65,6 +66,14 @@ body {
 
 .wrapper {
   border: 1px solid #7CA3EE;
+  padding: 0.6em;
+  display: grid;
+  grid-template-columns: 1fr;
+  row-gap: 1em;
+  margin-top: 1em;
+}
+
+.game-wrapper {
   padding: 0.6em;
   display: grid;
   grid-template-columns: 1fr;
@@ -124,14 +133,11 @@ body {
 }
 
 .word-grid > button {
-  background-color: #eee;
+  background-color: #8AA9E4;
   margin: 1em;
   padding: 1em;
   border: 1px solid darkgrey;
   font-size: 1em;
-}
-.word-grid > button:nth-child(odd) {
-  background-color: #ddd;
 }
 
 .task > p {
@@ -148,7 +154,7 @@ body {
 }
 .taskPBorder{
   border: 1px solid darkgrey;
-  background-color: #eee;
+  background-color: #8AA9E4;
   padding: .3em;
   min-width: 2em;
   min-height: 1.2em;
